Add OpenCV flip benchmark
